Last fixes to meet all rule requirements. Changed storybook to resolve JSX instead of JS
