+++ .eslintrc
@@ -8,6 +8,10 @@
   },
   "rules": {
     "prettier/prettier": ["error"],
-    "linebreak-style": 0
+    "linebreak-style": 0,
+    "import/no-extraneous-dependencies": ["error", {"devDependencies": true}]
+  },
+  "settings": {   
+     "import/resolver": "webpack"
   }
 }
