+++ package.json
@@ -5,7 +5,6 @@
   "main": "dist/index.js",
   "jsnext:main": "src/index.js",
   "scripts": {
-    "start": "webpack --watch",
     "test": "jest --coverage",
     "test:generate-output": "jest --json --outputFile=.jest-test-results.json || true",
     "prebuild:storybook": "yarn run test:generate-output",
@@ -49,6 +48,7 @@
     "eslint": "5.3.0",
     "eslint-config-airbnb": "17.1.0",
     "eslint-config-prettier": "^3.6.0",
+    "eslint-import-resolver-webpack": "^0.11.0",
     "eslint-loader": "^2.1.1",
     "eslint-plugin-import": "2.14.0",
     "eslint-plugin-jsx-a11y": "6.1.1",
