+++ src/components/feedback/Spinner/__tests__/Spinner.stories.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { storiesOf } from "@storybook/react";
+import Spinner from "..";
+
+storiesOf("Feedback components/Spinner", module)
+    .addParameters({ jest: ["Spinner.stories"] })
+    .add("Default", () => (
+        <React.Fragment>
+            <Spinner />
+            <Spinner primary />
+            <Spinner small />
+            <p>
+                Inline spinner <Spinner inline />
+            </p>
+        </React.Fragment>
+    ));
