+++ webpack.config.js
@@ -1,6 +1,8 @@
 const merge = require("webpack-merge");
 const styles = require("./build/style");
+const fonts = require("./build/fonts");
 const eslint = require("./build/eslint");
 const prettier = require("./build/prettier");
+const resolve = require("./build/resolve");
 
-module.exports = merge(styles, eslint, prettier, {});
+module.exports = merge(styles, fonts, resolve, eslint, prettier);
